requirements.txt
flask==3.0.0
flask-cors==4.0.0
websocket-client==1.7.0
pyotp==2.9.0
requests==2.31.0
